Change the declarations part of the grammar
Now all declarations in a compund statement must come before any
statement. Note that variables can only be declared in this part,
not defined.

This makes the task for the symbol manager easier as now all
variables are declared before each "chunk" of statements. Thus
the symbol manager does not need to handle variable declarations
mixed with statements.