Add generic_visit() to NodeVisitor
This new functions replaces the visit_Undefined function. The
generic_visit function visits all children of a node.

Now subclasses of the NodeVisitor no longer have to implement
visit functions for all types of nodes. For node that does not
have a visit implementation generic_visit() will be used to
visit the children of that node.

The generic_visit function can also be used in visit
implementations as a convenient way to visit all childrens of a
node.