Implement the symbol table
Implemented as a tree of hash tables where each node corresponds
to a scope. Two types of symbols are added: builtin types and
variables. More will be added later and the existing ones will
most likely be extended.

The symbol table is built via an implementation of the
NodeVisitor. Each node of the abstract syntax tree gets extended
with a symbol_table member that points to the corresponding node
of the symbol table.

The int and double types are automatically added to the global
scope of the symbol table upon creation.