Add vertification at creation of node visitors
Verify that there exists a corresponding node for all _visit_XXX
functions. This will catch spelling mistakes.

We would also want to catch when a visitor has been forgotten
but that is difficult since some node visitors do not need an
implementation for all nodes.

This implementation is a bit of a hack, there may be a better
way of doing this.