Add source code location to semantic errors
To be able to identify where in the code the error occured.

To enable this the token that lead to the creation of a node in
the abstract syntax tree is stored in the node. In cases where
multiple tokens were used to create a node the most "reasonable"
token is stored.