Improve support for functions
This changes includes:
* A translation unit node is added to the abstract syntax
  tree (AST), containing a list of functions.
* The parser creates translation unit nodes.
* Translation unit, function and parameter are now included
  in the AST graph.

Finally the symbol table is extended to handle functions,
including the parameters. Functions and the parameters are
individual entries in the symbol table. To keep track of
which parameters belong to which function each function
contains a list of its parameters.

The interpreter still lacks proper support for functions.
Function calls are not supported by any part of the
compiler, yet.