Add support for function calls to the sea sub compiler
The entire chain is added, from parser to optimizer
including the abstract syntax and symbol table. Note
however that the interpreter does not have support for
function calls yet.

The semantic analyzer verifies that all functions are
called with the correct number of arguments and that the
arguments are of the correct type.

The optimizer performs optimization on the arguments of the
function calls.