Prepare symbol table for target code generation
A function needs to know how many local variables it has
and their types in order to allocate the correct size on
the stack.

Since the intermediate code generator may add more local
variables to a function the function symbol must have
support for this.

For consistency the parameter handling of a function symbol
is changed to match the handling of local variables. This
also made the symbol table visitor a bit nicer.
order to allocate the correct size on the stack.