Remove the interpreter
The interpetator was added as a tool to be used in the
development of the compiler, it has now played its part.

It has not kept up with the development of the compiler and
there is really not any point in updating it, Sea sub is a
compiler, not an interpreter.